Extract media validation notices
