+++ projects/js-packages/publicize-components/src/components/form/index.tsx
@@ -8,30 +8,21 @@
 
 import { Disabled, PanelRow } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
-import { Fragment, useMemo } from '@wordpress/element';
+import { Fragment } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { usePublicizeConfig } from '../../..';
-import useAttachedMedia from '../../hooks/use-attached-media';
-import useDismissNotice from '../../hooks/use-dismiss-notice';
-import useFeaturedImage from '../../hooks/use-featured-image';
-import useImageGeneratorConfig from '../../hooks/use-image-generator-config';
-import useMediaDetails from '../../hooks/use-media-details';
-import useMediaRestrictions, { NO_MEDIA_ERROR } from '../../hooks/use-media-restrictions';
-import useRefreshAutoConversionSettings from '../../hooks/use-refresh-auto-conversion-settings';
 import useSocialMediaConnections from '../../hooks/use-social-media-connections';
 import { store as socialStore } from '../../social-store';
 import { ThemedConnectionsModal as ManageConnectionsModal } from '../manage-connections-modal';
 import { AdvancedPlanNudge } from './advanced-plan-nudge';
-import { AutoConversionNotice } from './auto-conversion-notice';
 import { BrokenConnectionsNotice } from './broken-connections-notice';
 import { ConnectionsList } from './connections-list';
-import { InstagramNoMediaNotice } from './instagram-no-media-notice';
+import { MediaValidationNotices } from './media-validation-notices';
 import { SettingsButton } from './settings-button';
 import { ShareCountInfo } from './share-count-info';
 import { SharePostForm } from './share-post-form';
 import styles from './styles.module.scss';
 import { UnsupportedConnectionsNotice } from './unsupported-connections-notice';
-import { ValidationNotice } from './validation-notice';
 
 /**
  * The Publicize form component. It contains the connection list, and the message box.
@@ -39,9 +30,7 @@ import { ValidationNotice } from './validation-notice';
  * @returns {object} - Publicize form component.
  */
 export default function PublicizeForm() {
-	const { connections, hasConnections, hasEnabledConnections } = useSocialMediaConnections();
-	const { isEnabled: isSocialImageGeneratorEnabledForPost } = useImageGeneratorConfig();
-	const { shouldShowNotice, NOTICES } = useDismissNotice();
+	const { hasConnections, hasEnabledConnections } = useSocialMediaConnections();
 	const {
 		isPublicizeEnabled,
 		isPublicizeDisabledBySitePlan,
@@ -49,48 +38,14 @@ export default function PublicizeForm() {
 		userConnectionUrl,
 	} = usePublicizeConfig();
 
-	const { numberOfSharesRemaining, useAdminUiV1 } = useSelect( select => {
+	const { useAdminUiV1 } = useSelect( select => {
 		const store = select( socialStore );
 		return {
-			numberOfSharesRemaining: store.numberOfSharesRemaining(),
 			useAdminUiV1: store.useAdminUiV1(),
 		};
 	}, [] );
 
 	const Wrapper = isPublicizeDisabledBySitePlan ? Disabled : Fragment;
-
-	const isAutoConversionEnabled = useSelect(
-		select => select( socialStore ).isAutoConversionEnabled(),
-		[]
-	);
-
-	const { attachedMedia } = useAttachedMedia();
-	const featuredImageId = useFeaturedImage();
-	const mediaId = attachedMedia[ 0 ]?.id || featuredImageId;
-
-	const { validationErrors, isConvertible } = useMediaRestrictions(
-		connections,
-		useMediaDetails( mediaId )[ 0 ],
-		{
-			isSocialImageGeneratorEnabledForPost,
-		}
-	);
-	const shouldAutoConvert = isAutoConversionEnabled && isConvertible;
-
-	const invalidIds = useMemo( () => Object.keys( validationErrors ), [ validationErrors ] );
-
-	const showValidationNotice = numberOfSharesRemaining !== 0 && invalidIds.length > 0;
-
-	const { refreshAutoConversionSettings } = useRefreshAutoConversionSettings();
-
-	if (
-		shouldAutoConvert &&
-		showValidationNotice &&
-		mediaId &&
-		shouldShowNotice( NOTICES.autoConversion )
-	) {
-		refreshAutoConversionSettings();
-	}
 
 	return (
 		<Wrapper>
@@ -106,17 +61,7 @@ export default function PublicizeForm() {
 					<ShareCountInfo />
 					<BrokenConnectionsNotice />
 					<UnsupportedConnectionsNotice />
-					{ shouldAutoConvert && showValidationNotice && mediaId && <AutoConversionNotice /> }
-					{ showValidationNotice &&
-						( Object.values( validationErrors ).includes( NO_MEDIA_ERROR ) ? (
-							<InstagramNoMediaNotice />
-						) : (
-							<ValidationNotice
-								connectionsCount={ connections.length }
-								invalidConnectionIdsCount={ invalidIds.length }
-								shouldAutoConvert={ shouldAutoConvert }
-							/>
-						) ) }
+					<MediaValidationNotices />
 				</>
 			) : null }
 			<PanelRow>
